adding Scheme to main API(traits.rs) & refactor
