+++ src/cyber/error.rs
@@ -3,18 +3,6 @@ use crate::backends::error::PointError;
 use crate::backends::error::ScalarError;
 
 #[derive(thiserror::Error, Debug)]
-pub enum SchemeError {
-    #[error("{0}")]
-    Ecies(#[from] EciesError),
-    #[error("{0}")]
-    Schnorr(#[from] SchnorrError),
-    #[error("{0}")]
-    Bls(#[from] BlsError),
-    #[error("{0}")]
-    TBls(#[from] TBlsError),
-}
-
-#[derive(thiserror::Error, Debug)]
 pub enum SchnorrError {
     #[error("invalid input lenght: expected {expected}, received {received}")]
     InvalidInputLenght { expected: usize, received: usize },
