--- src/drand/mod.rs
@@ -1,8 +1,0 @@
-pub mod ecies;
-pub mod poly;
-pub mod scheme;
-pub mod schnorr;
-pub mod tbls;
-
-pub mod error;
-pub use self::scheme::{BeaconDigest, DefaultScheme, Scheme, SchortSigScheme, UnchainedScheme};
